+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## dev
+
+## 0.1.0-b1
+- [feature] remember navigation [#2](https://github.com/Benature/obsidian-default-query-in-backlink/issues/2)
+- [feature] default collapse, extra context, sort order
+
 ## 0.0.2
 - [updated] support hot rewrite query when settings is changed
 - [feature] auto trigger query